+++ R/indicatorTreatment.R
@@ -11,6 +11,42 @@
   vals
 }
 
+as_rquery.vtreat_cat_ind <- function(tstep, 
+                                     ...,
+                                     var_restriction) {
+  if(!requireNamespace("rquery", quietly = TRUE)) {
+    stop("vtreat::as_rquery.vtreat_cat_ind treatmentplan requires the rquery package")
+  }
+  wrapr::stop_if_dot_args(substitute(list(...)), "vtreat::as_rquery.vtreat_cat_ind")
+  if((!is.null(var_restriction)) && (length(intersect(tstep$newvars, var_restriction))<=0)) {
+    return(NULL)
+  }
+  origvar <- tstep$origvar
+  exprs <- c()
+  for(i in seq_len(length(tstep$arg$tracked))) {
+    li <- tstep$arg$tracked[[i]]
+    vi <- tstep$newvars[[i]]
+    if(is.null(var_restriction) || (vi %in% var_restriction)) {
+      if(li == "NA") {
+        expri <- paste0("ifelse(is.na(", origvar, "), 1, 0)")
+      } else {
+        li <- gsub("^x ", "", li)
+        expri <- paste0("ifelse(is.na(", origvar, "), 0, ifelse(", origvar, " == \"", li, "\", 1, 0))")
+      }
+      names(expri) <- vi
+      exprs <- c(exprs, expri)
+    }
+  }
+  if(length(exprs)<=0) {
+    return(NULL)
+  }
+  list(
+    exprs = exprs,
+    optree_generators = list(),
+    tables = list()
+  )
+}
+
 # same signature as .mkCatInd (except no parallelCluster argument)
 .mkCatInd_a <- function(origVarName,
                         vcolin,
@@ -22,7 +58,7 @@
   if(length(tracked)<=0) {
     return(c())
   }
-  newVarNames <- make.names(paste(origVarName,'lev',tracked,sep="_"),unique=TRUE)
+  newVarNames <- vtreat_make_names(paste(origVarName,'lev',tracked,sep="_"))
   treatment <- list(origvar=origVarName,
                     newvars=newVarNames,
                     f=.catInd,
@@ -32,7 +68,7 @@
                     treatmentCode='lev',
                     needsSplit=FALSE,
                     extraModelDegrees=0)
-  class(treatment) <- 'vtreatment'
+  class(treatment) <- c('vtreat_cat_ind', 'vtreatment')
   pred <- treatment$f(vcolin,treatment$args)
   treatment$pred <- pred
   treatment