+++ R/prevalenceFact.R
@@ -14,6 +14,28 @@
   pred
 }
 
+
+as_rquery.vtreat_cat_p <- function(tstep, 
+                                     ...,
+                                     var_restriction) {
+  if(!requireNamespace("rquery", quietly = TRUE)) {
+    stop("vtreat::as_rquery.vtreat_cat_p treatmentplan requires the rquery package")
+  }
+  wrapr::stop_if_dot_args(substitute(list(...)), "vtreat::as_rquery.vtreat_cat_p")
+  if((!is.null(var_restriction)) && (!(tstep$newvars %in% var_restriction))) {
+    return(NULL)
+  }
+  args <- tstep$args
+  rquery_code_categorical(colname = tstep$origvar, 
+                          resname = tstep$newvars,
+                          coding_levels = names(args$scores),
+                          effect_values = args$scores,
+                          levRestriction = args$levRestriction,
+                          default_value = 0.0)
+}
+
+
+
 # build a prevalence fact
 .mkCatP <- function(origVarName,vcolin,zoY,zC,zTarget,levRestriction,weights,catScaling) {
   vcol <- .preProcCat(vcolin,c())
@@ -22,7 +44,7 @@
   den <- sum(weights)
   scores <- num/den
   scores <- as.list(scores)
-  newVarName <- make.names(paste(origVarName,'catP',sep='_'))
+  newVarName <- vtreat_make_names(paste(origVarName,'catP',sep='_'))
   treatment <- list(origvar=origVarName,
                     newvars=newVarName,
                     f=.catP,
@@ -36,7 +58,7 @@
   if(!.has.range.cn(pred)) {
     return(NULL)
   }
-  class(treatment) <- 'vtreatment'
+  class(treatment) <- c('vtreat_cat_p','vtreatment')
   if((!catScaling)||(is.null(zC))) {
     treatment$scales <- linScore(newVarName,pred,zoY,weights)
   } else {