+++ R/vtreat.R
@@ -51,12 +51,19 @@ vnames <- function(x) { x$newvars }
 #' @param x treatment plan
 #' @param ... additional args (to match general signature).
 #' @export
-format.vtreatment <- function(x,...) { paste(
-  'vtreat \'',x$treatmentName,
-  '\'(\'',x$origvar,'\'(',x$origType,',',x$origClass,')->',
-  x$convertedColClass,'->\'',
-  paste(x$newvars,collapse='\',\''),
-  '\')',sep='') }
+format.vtreatment <- function(x, ...) { 
+  paste(
+    'vtreat \'',x$treatmentName,
+    '\'(\'',x$origvar,'\'(',x$origType,',',x$origClass,')->',
+    x$convertedColClass,'->\'',
+    paste(x$newvars,collapse='\',\''),
+    '\')',sep='') 
+}
+
+#' @export
+as.character.vtreatment <- function (x, ...) {
+  format(x, ...)
+}
 
 #'
 #' Print treatmentplan.
@@ -64,10 +71,33 @@ format.vtreatment <- function(x,...) { paste(
 #' @param ... additional args (to match general signature).
 #' @seealso \code{\link{designTreatmentsC}} \code{\link{designTreatmentsN}} \code{\link{designTreatmentsZ}} \code{\link{prepare}}
 #' @export
-print.vtreatment <- function(x,...) { 
-  print(format.vtreatment(x),...) 
+print.vtreatment <- function(x, ...) { 
+  print(format(x), ...) 
 }
 
+
+
+
+
+#' @export
+format.treatmentplan <- function(x, ...) { 
+  format(x$scoreFrame)
+}
+
+#' @export
+as.character.treatmentplan <- function (x, ...) {
+  format(x, ...)
+}
+
+#'
+#' Print treatmentplan.
+#' @param x treatmentplan
+#' @param ... additional args (to match general signature).
+#' @seealso \code{\link{designTreatmentsC}} \code{\link{designTreatmentsN}} \code{\link{designTreatmentsZ}} \code{\link{prepare}}
+#' @export
+print.treatmentplan <- function(x, ...) { 
+  print(format(x), ...) 
+}
 
 
 
@@ -138,8 +168,8 @@ designTreatmentsC <- function(dframe,varlist,outcomename,outcometarget,
                               catScaling=FALSE,
                               verbose=TRUE,
                               parallelCluster=NULL) {
-  
-  .checkArgs(dframe=dframe,varlist=varlist,outcomename=outcomename,...)
+  wrapr::stop_if_dot_args(substitute(list(...)), "vtreat::designTreatmentsC")
+  .checkArgs(dframe=dframe,varlist=varlist,outcomename=outcomename)
   if(!(outcomename %in% colnames(dframe))) {
     stop("outcomename must be a column name of dframe")
   }
@@ -230,7 +260,8 @@ designTreatmentsN <- function(dframe,varlist,outcomename,
                               forceSplit=FALSE,
                               verbose=TRUE,
                               parallelCluster=NULL) {
-  .checkArgs(dframe=dframe,varlist=varlist,outcomename=outcomename,...)
+  wrapr::stop_if_dot_args(substitute(list(...)), "vtreat::designTreatmentsN")
+  .checkArgs(dframe=dframe,varlist=varlist,outcomename=outcomename)
   if(!(outcomename %in% colnames(dframe))) {
     stop("outcomename must be a column name of dframe")
   }
@@ -308,6 +339,7 @@ designTreatmentsZ <- function(dframe,varlist,
                               customCoders=NULL,
                               verbose=TRUE,
                               parallelCluster=NULL) {
+  wrapr::stop_if_dot_args(substitute(list(...)), "vtreat::designTreatmentsZ")
   # build a name disjoint from column names
   outcomename <- setdiff(paste('VTREATTEMPCOL',
                                seq_len(ncol(dframe) + length(varlist) + 1), 
@@ -315,7 +347,7 @@ designTreatmentsZ <- function(dframe,varlist,
                          c(colnames(dframe),varlist))[[1]]
   catScaling <- FALSE
   dframe[[outcomename]] <- 0
-  .checkArgs(dframe=dframe,varlist=varlist,outcomename=outcomename,...)
+  .checkArgs(dframe=dframe,varlist=varlist,outcomename=outcomename)
   ycol <- dframe[[outcomename]]
   treatments <- .designTreatmentsX(dframe,varlist,outcomename,ycol,
                      c(),c(),
@@ -484,7 +516,8 @@ prepare <- function(treatmentplan, dframe,
                     codeRestriction= NULL,
                     trackedValues= NULL,
                     parallelCluster= NULL) {
-  .checkArgs1(dframe=dframe,...)
+  wrapr::stop_if_dot_args(substitute(list(...)), "vtreat::prepare")
+  .checkArgs1(dframe=dframe)
   if(class(treatmentplan)!='treatmentplan') {
     stop("treatmentplan must be of class treatmentplan")
   }
@@ -630,7 +663,8 @@ mkCrossFrameCExperiment <- function(dframe,varlist,
                                     catScaling=FALSE,
                                     verbose= TRUE,
                                     parallelCluster=NULL) {
-  .checkArgs(dframe=dframe,varlist=varlist,outcomename=outcomename,...)
+  wrapr::stop_if_dot_args(substitute(list(...)), "vtreat::mkCrossFrameCExperiment")
+  .checkArgs(dframe=dframe,varlist=varlist,outcomename=outcomename)
   if(!is.data.frame(dframe)) {
     stop("dframe must be a data frame")
   }
@@ -769,7 +803,8 @@ mkCrossFrameNExperiment <- function(dframe,varlist,outcomename,
                                     forceSplit=FALSE,
                                     verbose= TRUE,
                                     parallelCluster=NULL) {
-  .checkArgs(dframe=dframe,varlist=varlist,outcomename=outcomename,...)
+  wrapr::stop_if_dot_args(substitute(list(...)), "vtreat::mkCrossFrameNExperiment")
+  .checkArgs(dframe=dframe,varlist=varlist,outcomename=outcomename)
   if(!is.data.frame(dframe)) {
     stop("dframe must be a data frame")
   }